% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.Pairwise.SpeciesTree.Distances.R
\name{Compute.Pairwise.SpeciesTree.Distances}
\alias{Compute.Pairwise.SpeciesTree.Distances}
\title{Compute.Pairwise.SpeciesTree.Distances: function to XXX}
\usage{
Compute.Pairwise.SpeciesTree.Distances(list.SpeciesTrees,
  vector.SpeciesTreeNames, string.PathParentDir, string.PathHybridCoal)
}
\arguments{
\item{XXX}{XXX}
}
\value{
vector.Probabilistic.SpeciesTree.Distances Vector containing the three PSTD measures\cr \cr
}
\description{
This function returns a XXX
}
\examples{

}
\keyword{coalescent,}
\keyword{distance,}
\keyword{hybridization}
\keyword{incomplete}
\keyword{lineage}
\keyword{model}
\keyword{multispecies}
\keyword{sorting,}
