% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.Probabilistic.SpeciesTree.Distances.R
\name{Compute.Probabilistic.SpeciesTree.Distances}
\alias{Compute.Probabilistic.SpeciesTree.Distances}
\title{Compute.Probabilistic.SpeciesTree.Distances: function to compute probabilistic species tree distances between two species tree models}
\usage{
Compute.Probabilistic.SpeciesTree.Distances(handle.SpeciesTree.Model1,
  handle.SpeciesTree.Model2, string.PathParentDir, string.PathHybridCoal)
}
\arguments{
\item{handle.SpeciesTree.Model1}{Species Tree Model 1}

\item{handle.SpeciesTree.Model2}{Species Tree Model 2}

\item{string.PathParentDir}{Path to a parent directory that will be used for subdirectories}

\item{string.PathHybridCoal}{Path to HYBRID-COAL executable}
}
\value{
vector.Probabilistic.SpeciesTree.Distances Vector containing the three PSTD measures\cr \cr
}
\description{
This function returns a vector of length (3): representing the "Hellinger", "Kullback_Leibler", and "Jensen_Shannon" distances between to trees
}
\examples{

string.SpeciesTree.1a <- "(((A:1,B:1):1,C:2):1,D:3);"
handle.SpeciesTree.1a <- read.tree(text = string.SpeciesTree.1a)

string.SpeciesTree.1b <- "(((A:1,C:1):1,B:2):1,D:3);"
handle.SpeciesTree.1b <- read.tree(text = string.SpeciesTree.1b)


Compute.Probabilistic.SpeciesTree.Distances(handle.SpeciesTree.Model1 = handle.SpeciesTree.1a, 
   handle.SpeciesTree.Model2 = handle.SpeciesTree.1b, 
   string.PathParentDir = '~/Desktop/EXAMPLE/SmallTree/', 
   string.PathHybridCoal = '/Applications/hybrid-coal-v0.2.1-beta/hybrid-coal')
}
\keyword{coalescent,}
\keyword{distance,}
\keyword{hybridization}
\keyword{incomplete}
\keyword{lineage}
\keyword{model}
\keyword{multispecies}
\keyword{sorting,}
