% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputePSTD}
\alias{ComputePSTD}
\title{ComputeDistancesPSTD}
\usage{
ComputePSTD(vectorM1, vectorM2)
}
\arguments{
\item{vectorM1}{Vector of the gene tree probabilities for M1}

\item{vectorM2}{Vector of the gene tree probabilities for M2}
}
\description{
Function to compute the probabililistic distance between two gene tree probabilities for the Hellinger (dH, first), KL (dKL, second), and JS (dJS, third) distance measures of Garba et al 2018

IMPORTANT: This function assumes that the two gene tree probability vectors (vectorM1 and vectorM2) are sorted by to be the same. That is, the first gene tree in vectorM1 is also the first gene tree in vectorM2, and so on...
The gene tree probability output files of HYBRID-COAL are sorted automatically.
}
\details{
\code{ComputeDistancesPSTD} Computes the probabilistic distances between two probability distributions
}
\examples{

vector.GeneTreeProbs.M1 <- Read.GeneTree.Probabilities(string.PathToGeneTreeProbs = '~/Desktop/EXAMPLE/SmallTree/M1.prob')
vector.GeneTreeProbs.M2 <- Read.GeneTree.Probabilities(string.PathToGeneTreeProbs = '~/Desktop/EXAMPLE/SmallTree/M2.prob')

ComputePSTD(vectorM1 = vector.GeneTreeProbs.M1, vectorM2 = vector.GeneTreeProbs.M2)

}
