% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pipeline.SpeciesTreeDistances.ScaleAllBranches.R
\name{Pipeline.SpeciesTreeDistances.ScaleAllBranches}
\alias{Pipeline.SpeciesTreeDistances.ScaleAllBranches}
\title{Pipeline.SpeciesTreeDistances.ScaleAllBranches: function to measure phylogenetic distances across a range of branch length scaling between two species tree models}
\usage{
Pipeline.SpeciesTreeDistances.ScaleAllBranches(vector.ScaleFactor,
  string.PathParentDir, string.PathHybridCoal, handle.SpeciesTree.Model1,
  handle.SpeciesTree.Model2)
}
\arguments{
\item{vector.ScaleFactor}{Vector to containing the values to scale the branches of the second species tree model}

\item{string.PathParentDir}{Path to a parent directory that will be used for subdirectories}

\item{string.PathHybridCoal}{Path to HYBRID-COAL executable}

\item{handle.SpeciesTree.Model1}{Species Tree Model 1}

\item{handle.SpeciesTree.Model2}{Species Tree Model 2}
}
\value{
XXX\cr \cr
}
\description{
This function returns a matrix containing the 5 distance metics (3 probabilistic + Robinson-Foulds + BHV)
}
\examples{

library(phangorn)
library(PSTD)
library(distory)

string.SpeciesTree.1a <- "(((A:1,B:1):1,C:2):1,D:3);"
handle.SpeciesTree.1a <- read.tree(text = string.SpeciesTree.1a)

vector.Gamma <- seq(0.01, 10, 0.01)
handle.Figure1d <- Pipeline.SpeciesTreeDistances.ScaleAllBranches(vector.ScaleFactor = vector.Gamma, 
   string.PathParentDir = '~/Desktop/', 
   string.PathHybridCoal = '/Applications/hybrid-coal-v0.2.1-beta/hybrid-coal', 
   handle.SpeciesTree.Model1 = handle.SpeciesTree.1a, 
   handle.SpeciesTree.Model2 = handle.SpeciesTree.1a)
   
plot(handle.Figure1d[,1], ylim = c(0, 0.75), col = "darkblue", type = "l", xaxt='n', yaxt="n", lty = 1)
axis(1,at=seq(from = 0, to = 1000, by = 50),labels=seq(from = 0, to = 10, by = 0.50))
axis(2,at=seq(from = 0, to = 0.75, by = 0.10),labels=seq(from = 0, to = 0.75, by = 0.10))
lines(handle.Figure1d[,2], col = "blue", lty = 2)
lines(handle.Figure1d[,3], col = "lightblue", lty = 4)
lines(handle.Figure1d[,5]*0.025, col = "red", lty = 3)
}
\keyword{coalescent,}
\keyword{distance,}
\keyword{hybridization}
\keyword{incomplete}
\keyword{lineage}
\keyword{model}
\keyword{multispecies}
\keyword{sorting,}
