% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pipeline.SpeciesTreeDistances.ScaleSingleBranch.R
\name{Pipeline.SpeciesTreeDistances.ScaleSingleBranch}
\alias{Pipeline.SpeciesTreeDistances.ScaleSingleBranch}
\title{Pipeline.SpeciesTreeDistances.ScaleSingleBranch: function to measure phylogenetic distances across a range of branch length scaling (for a single branch on the second model) between two species tree models}
\usage{
Pipeline.SpeciesTreeDistances.ScaleSingleBranch(vector.ScaleFactor,
  string.PathParentDir, string.PathHybridCoal, handle.SpeciesTree.Model1,
  handle.SpeciesTree.Model2, numeric.BranchToScale)
}
\arguments{
\item{vector.ScaleFactor}{Vector to containing the values to scale the branches of the second species tree model}

\item{string.PathParentDir}{Path to a parent directory that will be used for subdirectories}

\item{string.PathHybridCoal}{Path to HYBRID-COAL executable}

\item{handle.SpeciesTree.Model1}{Species Tree Model 1}

\item{handle.SpeciesTree.Model2}{Species Tree Model 2}

\item{numeric.BranchToScale}{Number indicating the branch to scale on species tree model 2}
}
\value{
XXX\cr \cr
}
\description{
This function returns a matrix containing the 5 distance metics (3 probabilistic + Robinson-Foulds + BHV) across a range of scaler values
}
\examples{

}
\keyword{coalescent,}
\keyword{distance,}
\keyword{hybridization}
\keyword{incomplete}
\keyword{lineage}
\keyword{model}
\keyword{multispecies}
\keyword{sorting,}
