% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.GeneTree.Probabilities.R
\name{Read.GeneTree.Probabilities}
\alias{Read.GeneTree.Probabilities}
\title{Read.GeneTree.Probabilities: function to extract the gene tree probabilities from the output of HYBRID-COAL}
\usage{
Read.GeneTree.Probabilities(string.PathToGeneTreeProbs)
}
\arguments{
\item{string.PathToGeneTreeProbs}{Vector containing a set of likelihoods under model 1}
}
\value{
vector.GeneTreeProbabilities: vector containing a set of gene tree probabilities\cr
}
\description{
This function returns a vector of containing gene tree probabilities (assumed to be in the same order when comparing two gene tree probability vectors)
}
\examples{
vector.GeneTreeProbs.M1 <- Read.GeneTree.Probabilities(string.PathToGeneTreeProbs = '~/Desktop/EXAMPLE/SmallTree/M1.prob')

}
\keyword{coalescent,}
\keyword{distance,}
\keyword{hybridization}
\keyword{incomplete}
\keyword{lineage}
\keyword{model}
\keyword{multispecies}
\keyword{sorting,}
